+++ Tests/Unit/Testing/Fixtures/TestingSubclass.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+namespace OliverKlee\Oelib\Tests\Unit\Testing\Fixtures;
+
+final class TestingSubclass extends \stdClass
+{
+}