+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 ## x.y.z
 
 ### Added
+- Add a helper for clearing the `GU::makeInstance()` class name cache (#1091)
 - Allow providing data for `TestingFramework::createFrontEndPage()` (#1086)
 
 ### Changed